[mqtt-http-to-s3-java] Migrate akka dependencies to pekko
